requirements.txt
aiogram>=3.7,<4.0
asyncpg>=0.29
redis>=5.0
python-dotenv>=1.0
